annex-d: UMV+PLUSPTYPE motion vectors switch to Table D.3, both directions
Closes the round-443 self-found nonconformance: §5.3.7 sends UMV
motion vector data with the Table D.3 reversible codes when PLUSPTYPE
is present, not Table 14. The parser reads MVD/MVD2-4/MVDB as §D.3
pairs (with the §D.2 six-zero emulation-prevention bit), the
reconstruction is the single-valued predictor + difference bounded by
the §5.1.9 UUI selection (Tables D.1/D.2, or unlimited — previously
refused), the last-sent UUI inherits per §5.1.4.4, and the encoder
emits §D.3 pairs under a Tables-D.1/D.2 + §D.1.1 search window.
Also corrects the §P.2.2 warping-parameter emulation-prevention
condition (value-+1 pairs, not zero-value pairs) and widens Mvd to
i16 for the §D.3 difference range.

Validated both directions: the staged UMV+ conformance fixture
decodes with zero samples beyond the Annex A.7 ±1 bound, and a
crate-encoded UMV+ stream black-box-decodes externally at max diff 1.

+++ src/pb_layer.rs
@@ -476,8 +476,8 @@ pub fn cbpb_block_present(cbpb: u8, block_number: u32) -> bool {
 /// (the read MVD VLC machinery scans the full 13-bit codeword domain
 /// before giving up).
 pub fn parse_mvdb(reader: &mut BitReader<'_>) -> Result<Mvd> {
-    let dx_half = decode_mvd_component(reader)?;
-    let dy_half = decode_mvd_component(reader)?;
+    let dx_half = decode_mvd_component(reader)? as i16;
+    let dy_half = decode_mvd_component(reader)? as i16;
     Ok(Mvd { dx_half, dy_half })
 }
